+++ tests/test_client.lua
@@ -1,39 +1,10 @@
 local Path = require "obsidian.path"
 local Note = require "obsidian.note"
-local obsidian = require "obsidian"
-
----Get a client in a temporary directory.
----
----@param run fun(client: obsidian.Client)
-local with_tmp_client = function(run)
-  local dir = Path.temp { suffix = "-obsidian" }
-  dir:mkdir { parents = true }
-
-  local client = obsidian.new_from_dir(tostring(dir))
-  client.opts.note_id_func = function(title)
-    local id = ""
-    if title ~= nil then
-      id = title:gsub(" ", "-"):gsub("[^A-Za-z0-9-]", ""):lower()
-    else
-      for _ = 1, 4 do
-        id = id .. string.char(math.random(65, 90))
-      end
-    end
-    return id
-  end
-
-  local ok, err = pcall(run, client)
-
-  dir:rmtree()
-
-  if not ok then
-    error(err)
-  end
-end
+local h = dofile "tests/helpers.lua"
 
 describe("Client", function()
   it("should be able to initialize a daily note", function()
-    with_tmp_client(function(client)
+    h.with_tmp_client(function(client)
       local note = client:today()
       MiniTest.expect.equality(true, note.path ~= nil)
       MiniTest.expect.equality(true, note:exists())
@@ -41,7 +12,7 @@ describe("Client", function()
   end)
 
   it("should not add frontmatter for today when disabled", function()
-    with_tmp_client(function(client)
+    h.with_tmp_client(function(client)
       client.opts.disable_frontmatter = true
       local new_note = client:today()
 
@@ -51,7 +22,7 @@ describe("Client", function()
   end)
 
   it("should not add frontmatter for yesterday when disabled", function()
-    with_tmp_client(function(client)
+    h.with_tmp_client(function(client)
       client.opts.disable_frontmatter = true
       local new_note = client:yesterday()
 
@@ -63,7 +34,7 @@ end)
 
 describe("Client:new_note_path()", function()
   it('should only append one ".md" at the end of the path', function()
-    with_tmp_client(function(client)
+    h.with_tmp_client(function(client)
       client.opts.note_path_func = function(spec)
         return (spec.dir / "foo-bar-123"):with_suffix ".md.md.md"
       end
@@ -77,7 +48,7 @@ end)
 
 describe("Client:parse_title_id_path()", function()
   it("should parse a title that's a partial path and generate new ID", function()
-    with_tmp_client(function(client)
+    h.with_tmp_client(function(client)
       local title, id, path = client:parse_title_id_path "notes/Foo"
       MiniTest.expect.equality("Foo", title)
       MiniTest.expect.equality("foo", id)
@@ -91,7 +62,7 @@ describe("Client:parse_title_id_path()", function()
   end)
 
   it("should interpret relative directories relative to vault root.", function()
-    with_tmp_client(function(client)
+    h.with_tmp_client(function(client)
       local title, id, path = client:parse_title_id_path("Foo", nil, "new-notes")
       MiniTest.expect.equality(title, "Foo")
       MiniTest.expect.equality(id, "foo")
@@ -100,7 +71,7 @@ describe("Client:parse_title_id_path()", function()
   end)
 
   it("should parse an ID that's a path", function()
-    with_tmp_client(function(client)
+    h.with_tmp_client(function(client)
       local title, id, path = client:parse_title_id_path("Foo", "notes/1234-foo")
       MiniTest.expect.equality(title, "Foo")
       MiniTest.expect.equality(id, "1234-foo")
@@ -109,7 +80,7 @@ describe("Client:parse_title_id_path()", function()
   end)
 
   it("should parse a title that's an exact path", function()
-    with_tmp_client(function(client)
+    h.with_tmp_client(function(client)
       local title, id, path = client:parse_title_id_path "notes/foo.md"
       MiniTest.expect.equality(title, "foo")
       MiniTest.expect.equality(id, "foo")
@@ -118,7 +89,7 @@ describe("Client:parse_title_id_path()", function()
   end)
 
   it("should ignore boundary whitespace when parsing a title", function()
-    with_tmp_client(function(client)
+    h.with_tmp_client(function(client)
       local title, id, path = client:parse_title_id_path "notes/Foo  "
       MiniTest.expect.equality(title, "Foo")
       MiniTest.expect.equality(id, "foo")
@@ -127,7 +98,7 @@ describe("Client:parse_title_id_path()", function()
   end)
 
   it("should keep whitespace within a path when parsing a title", function()
-    with_tmp_client(function(client)
+    h.with_tmp_client(function(client)
       local title, id, path = client:parse_title_id_path "notes/Foo Bar.md"
       MiniTest.expect.equality(title, "Foo Bar")
       MiniTest.expect.equality(id, "Foo Bar")
@@ -136,7 +107,7 @@ describe("Client:parse_title_id_path()", function()
   end)
 
   it("should keep allow decimals in ID", function()
-    with_tmp_client(function(client)
+    h.with_tmp_client(function(client)
       local title, id, path = client:parse_title_id_path("Title", "johnny.decimal", "notes")
       MiniTest.expect.equality(title, "Title")
       MiniTest.expect.equality(id, "johnny.decimal")
@@ -145,7 +116,7 @@ describe("Client:parse_title_id_path()", function()
   end)
 
   it("should generate a new id when the title is just a folder", function()
-    with_tmp_client(function(client)
+    h.with_tmp_client(function(client)
       local title, id, path = client:parse_title_id_path "notes/"
       MiniTest.expect.equality(title, nil)
       MiniTest.expect.equality(#id, 4)
@@ -154,7 +125,7 @@ describe("Client:parse_title_id_path()", function()
   end)
 
   it("should respect configured 'note_path_func'", function()
-    with_tmp_client(function(client)
+    h.with_tmp_client(function(client)
       client.opts.note_path_func = function(spec)
         return (spec.dir / "foo-bar-123"):with_suffix ".md"
       end
@@ -167,7 +138,7 @@ describe("Client:parse_title_id_path()", function()
   end)
 
   it("should ensure result of 'note_path_func' always has '.md' suffix", function()
-    with_tmp_client(function(client)
+    h.with_tmp_client(function(client)
       client.opts.note_path_func = function(spec)
         return spec.dir / "foo-bar-123"
       end
@@ -180,7 +151,7 @@ describe("Client:parse_title_id_path()", function()
   end)
 
   it("should ensure result of 'note_path_func' is always an absolute path and within provided directory", function()
-    with_tmp_client(function(client)
+    h.with_tmp_client(function(client)
       client.opts.note_path_func = function(_)
         return "foo-bar-123.md"
       end
@@ -197,7 +168,7 @@ end)
 
 describe("Client:_prepare_search_opts()", function()
   it("should prepare search opts properly", function()
-    with_tmp_client(function(client)
+    h.with_tmp_client(function(client)
       ---@diagnostic disable-next-line: invisible
       local opts = client:_prepare_search_opts(true, { max_count_per_file = 1 })
       MiniTest.expect.equality(opts:to_ripgrep_opts(), { "--sortr=modified", "-m=1" })
@@ -207,14 +178,14 @@ end)
 
 describe("Client:vault_relative_path()", function()
   it("should resolve relative paths", function()
-    with_tmp_client(function(client)
+    h.with_tmp_client(function(client)
       MiniTest.expect.equality(client:vault_relative_path "foo.md", Path.new "foo.md")
       MiniTest.expect.equality(client:vault_relative_path(client.dir / "foo.md"), Path.new "foo.md")
     end)
   end)
 
   it("should error when strict=true and the relative path can't be resolved", function()
-    with_tmp_client(function(client)
+    h.with_tmp_client(function(client)
       MiniTest.expect.error(function()
         client:vault_relative_path("/Users/petew/foo.md", { strict = true })
       end)
@@ -222,7 +193,7 @@ describe("Client:vault_relative_path()", function()
   end)
 
   it("should not error when strict=false and the relative path can't be resolved", function()
-    with_tmp_client(function(client)
+    h.with_tmp_client(function(client)
       MiniTest.expect.equality(nil, client:vault_relative_path "/Users/petew/foo.md")
     end)
   end)
@@ -230,7 +201,7 @@ end)
 
 describe("Client:create_note()", function()
   it("should create a new note with or without aliases and tags", function()
-    with_tmp_client(function(client)
+    h.with_tmp_client(function(client)
       local note = client:create_note { title = "Foo", aliases = { "Bar" }, tags = { "note" } }
       MiniTest.expect.equality(note.title, "Foo")
       MiniTest.expect.equality(note.aliases, { "Bar", "Foo" })
@@ -242,7 +213,7 @@ end)
 
 describe("Client:daily_note_path()", function()
   it("should use the path stem as the ID", function()
-    with_tmp_client(function(client)
+    h.with_tmp_client(function(client)
       client.opts.daily_notes.date_format = "%Y/%b/%Y-%m-%d"
       local path, id = client:daily_note_path()
       assert(vim.endswith(tostring(path), tostring(os.date("%Y/%b/%Y-%m-%d.md", os.time()))))